Be explicit about rounding away from zero.
The _compute_percentage function uses the round function which changes
behavior between python 2 and python 3.  This could potentially change
the way grades are calculated.  To maintain the behavior we had
previously we just explicityl call math.floor after adding .5 to round
the floating point number away from zero.  This was the old behavior of
round for positive numbers and we just keep it that even in python 3.